Move the tombstone test into modelserver_controller_test.go
It exercises enqueueModelServer, so it belongs with the other tests for
that controller rather than in a file of its own.
